# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервиса для сокращения URL",
        "title": "URL Shortener API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "./cmd/url-shortener",
    "paths": {
        "/url": {
            "post": {
                "description": "Преобразует длинный URL в короткий",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Создать короткую ссылку",
                "parameters": [
                    {
                        "description": "URL для сокращения",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handlers_url_save.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handlers_url_save.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_m1al04949_url-shortener_internal_lib_api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_m1al04949_url-shortener_internal_lib_api_response.Response"
                        }
                    }
                }
            }
        },
        "/url/{alias}": {
            "delete": {
                "description": "Удаляет короткий URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Удалять короткую ссылку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Алиас короткой ссылки",
                        "name": "alias",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handlers_delete.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_m1al04949_url-shortener_internal_lib_api_response.Response"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_m1al04949_url-shortener_internal_lib_api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_m1al04949_url-shortener_internal_lib_api_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_m1al04949_url-shortener_internal_lib_api_response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_http-server_handlers_delete.Response": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "response": {
                    "$ref": "#/definitions/github_com_m1al04949_url-shortener_internal_lib_api_response.Response"
                }
            }
        },
        "internal_http-server_handlers_url_save.Request": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "alias": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "internal_http-server_handlers_url_save.Response": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}